if_else.java
This is a simple Java program that demonstrates the use of the if-else conditional statement. The program prompts the user to enter their age and then checks whether the user is eligible to get a driving license.

If the age is below 18, it prints that the user is not eligible.

This program is useful as a beginner-level example to understand how conditional logic works in Java.