PersonalDetails.java
This is a beginner-level Java program that demonstrates how to collect and display personal information using user input. It makes use of the Scanner class to read various types of data from the user, including:

Name (as a String)

Age (as an int)

Phone number (as a long)

Address (as a String)

After collecting the data, the program neatly prints the details entered by the user. This example is useful for understanding:

Input/output operations

Handling different data types

Managing newline issues when mixing nextInt()/nextLong() and nextLine()